backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── OCR & Document Processing ─────────────────────────────
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
python-docx==1.1.2
openpyxl==3.1.2

# ── Data & ML ─────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
mlflow==2.12.2

# ── LLM Integration ───────────────────────────────────────
google-generativeai==0.6.0

# ── Security & Encryption ─────────────────────────────────
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Utilities ─────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1

# ── Monitoring & Logging ──────────────────────────────────
structlog==24.1.0
prometheus-client==0.20.0

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ── Development ───────────────────────────────────────────
watchdog==4.0.0
